Mark additional tests for running with single-threaded dask
